add valid, mapper

+++ src/main/java/vdtry06/springboot/authservice/service/UserService.java
@@ -7,59 +7,61 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import vdtry06.springboot.authservice.dto.request.UserCreationRequest;
 import vdtry06.springboot.authservice.dto.request.UserUpdationRequest;
+import vdtry06.springboot.authservice.dto.response.UserResponse;
 import vdtry06.springboot.authservice.entity.User;
 import vdtry06.springboot.authservice.exception.AppException;
 import vdtry06.springboot.authservice.exception.ErrorCode;
+import vdtry06.springboot.authservice.mapper.UserMapper;
 import vdtry06.springboot.authservice.repository.UserRepository;
 
 import java.util.List;
 
+import static vdtry06.springboot.authservice.exception.ErrorCode.USER_NOT_EXISTED;
+
+@Slf4j
 @Service
 @RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
-@Slf4j
 public class UserService {
     UserRepository userRepository;
+    UserMapper userMapper;
 
-    public User createUser(UserCreationRequest request) {
-        User user = new User();
-
+    public UserResponse createUser(UserCreationRequest request) {
         if (userRepository.existsByUsername(request.getUsername()))
             throw new AppException(ErrorCode.USER_EXISTED);
 
-        user.setUsername(request.getUsername());
-        user.setPassword(request.getPassword());
-        user.setEmail(request.getEmail());
-        user.setFirstName(request.getFirstName());
-        user.setLastName(request.getLastName());
-        user.setDateOfBirth(request.getDateOfBirth());
+        User user = userMapper.toUser(request);
+
         log.info("User created: " + user);
-        return userRepository.save(user);
+        return userMapper.toUserResponse(userRepository.save(user));
     }
 
-    public User updateUser(String userId, UserUpdationRequest request) {
-        User user = getUser(userId);
+    public UserResponse updateUser(String userId, UserUpdationRequest request) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTED));
 
-        user.setPassword(request.getPassword());
-        user.setEmail(request.getEmail());
-        user.setFirstName(request.getFirstName());
-        user.setLastName(request.getLastName());
-        user.setDateOfBirth(request.getDateOfBirth());
+        userMapper.updateUser(user, request);
+
         log.info("Updating user with id: " + userId);
-        return userRepository.save(user);
+        return userMapper.toUserResponse(userRepository.save(user));
     }
 
-    public List<User> getUsers() {
+    public List<UserResponse> getUsers() {
         log.info("Getting all users");
-        return userRepository.findAll();
+        return userRepository.findAll().stream().map(userMapper::toUserResponse).toList();
     }
 
-    public User getUser(String id) {
+    public UserResponse getUser(String id) {
         log.info("Getting user with id: " + id);
-        return userRepository.findById(id).orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTED));
+        return userMapper.toUserResponse(userRepository.findById(id).orElseThrow(() -> new AppException(USER_NOT_EXISTED)));
     }
 
     public void deleteUser(String id) {
+        if (!userRepository.existsById(id)) {
+            log.info("User not existed!");
+            throw new AppException(ErrorCode.USER_NOT_EXISTED);
+        }
+
         log.info("Deleting user {}", id);
         userRepository.deleteById(id);
     }
